Address code review feedback for branch name display
- Extract worktree_info_from_thread_paths into shared function in
  thread_metadata_store.rs, fixing missing disambiguation logic in the
  archive view
- Unify tooltip to show individual full_path per chip instead of
  aggregate paths for linked worktrees
- Restructure worktree chip rendering loop to use match, removing dead
  code path
- Compute has_worktree from worktree_labels after the loop instead of
  predicting it eagerly
- Prune stale branch_names in update_worktree_paths when paths change
- Add text truncation with ellipsis to worktree chips, preferring to
  truncate worktree name over branch name
