+++ crates/agent_ui/src/threads_archive_view.rs
@@ -3,7 +3,9 @@ use std::sync::Arc;
 
 use crate::agent_connection_store::AgentConnectionStore;
 
-use crate::thread_metadata_store::{ThreadId, ThreadMetadata, ThreadMetadataStore};
+use crate::thread_metadata_store::{
+    ThreadId, ThreadMetadata, ThreadMetadataStore, worktree_info_from_thread_paths,
+};
 use crate::{Agent, DEFAULT_THREAD_TITLE, RemoveSelectedThread};
 
 use agent::ThreadStore;
@@ -23,10 +25,10 @@ use picker::{
     Picker, PickerDelegate,
     highlighted_match_with_paths::{HighlightedMatch, HighlightedMatchWithPaths},
 };
-use project::{AgentId, AgentServerStore, linked_worktree_short_name};
+use project::{AgentId, AgentServerStore};
 use settings::Settings as _;
 use theme::ActiveTheme;
-use ui::{AgentThreadStatus, ThreadItem, ThreadItemWorktreeInfo, WorktreeKind};
+use ui::{AgentThreadStatus, ThreadItem};
 use ui::{
     Divider, KeyBinding, ListItem, ListItemSpacing, ListSubHeader, Tooltip, WithScrollbar,
     prelude::*, utils::platform_title_bar_height,
@@ -537,34 +539,8 @@ impl ThreadsArchiveView {
 
                 let is_restoring = self.restoring.contains(&thread.thread_id);
 
-                let worktrees: Vec<ThreadItemWorktreeInfo> = thread
-                    .worktree_paths
-                    .ordered_pairs()
-                    .filter_map(|(main_path, folder_path)| {
-                        let is_linked = main_path != folder_path;
-                        let branch_name = thread.branch_names.get(folder_path).cloned();
-                        if is_linked {
-                            let name = linked_worktree_short_name(main_path, folder_path)
-                                .unwrap_or_default();
-                            Some(ThreadItemWorktreeInfo {
-                                name,
-                                full_path: SharedString::from(folder_path.display().to_string()),
-                                highlight_positions: Vec::new(),
-                                kind: WorktreeKind::Linked,
-                                branch_name,
-                            })
-                        } else {
-                            let name = folder_path.file_name()?;
-                            Some(ThreadItemWorktreeInfo {
-                                name: SharedString::from(name.to_string_lossy().to_string()),
-                                full_path: SharedString::from(folder_path.display().to_string()),
-                                highlight_positions: Vec::new(),
-                                kind: WorktreeKind::Main,
-                                branch_name,
-                            })
-                        }
-                    })
-                    .collect();
+                let worktrees =
+                    worktree_info_from_thread_paths(&thread.worktree_paths, &thread.branch_names);
 
                 let base = ThreadItem::new(id, thread.display_title())
                     .icon(icon)
